% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_list_columns_tweets.R
\name{combine_list_columns_tweets}
\alias{combine_list_columns_tweets}
\title{This function is a sub function combining listed columns into a string for simpler handling in a data frame}
\usage{
combine_list_columns_tweets(h, dt)
}
\arguments{
\item{h}{string containing the column name}

\item{dt}{data table or data frame}
}
\value{
a vector
}
\description{
This function is a sub function combining listed columns into a string for simpler handling in a data frame
}
\examples{
\dontrun{
}
}
