% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_json.R
\name{data_json}
\alias{data_json}
\title{This function is a sub function parsing the data returned from the API}
\usage{
data_json(data_twitter = data)
}
\arguments{
\item{data_twitter}{list with results from the API call}
}
\value{
list wirtable as a JSON file
}
\description{
This function is a sub function parsing the data returned from the API
}
\examples{
\dontrun{
}
}
