% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser_stream.R
\name{data_parser_stream}
\alias{data_parser_stream}
\title{This function is a sub function parsing the data returned from the API}
\usage{
data_parser_stream(results_data, filtered)
}
\arguments{
\item{results_data}{list with results from the API call}

\item{filtered}{logical, indicating which stream endpoint was used in the data (filtered or sampled)}
}
\value{
data frame
}
\description{
This function is a sub function parsing the data returned from the API
}
