% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parser_users.R
\name{data_parser_users}
\alias{data_parser_users}
\title{This function is a sub function parsing the data returned from the API}
\usage{
data_parser_users(results_data)
}
\arguments{
\item{results_data}{list with results from the API call}
}
\value{
data frame
}
\description{
This function is a sub function parsing the data returned from the API
}
