% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_stream_add_rule.R
\name{filtered_stream_add_rule}
\alias{filtered_stream_add_rule}
\title{This function allows you to collect tweets from the 1 \% sampled stream from twitter}
\usage{
filtered_stream_add_rule(token = "", value = "", tag = "")
}
\arguments{
\item{token}{string of the bearer token from your personal twitter API access}

\item{value}{string with rule value}

\item{tag}{string with rule tag}
}
\value{
a list
}
\description{
This function allows you to collect tweets from the 1 \% sampled stream from twitter
}
\examples{
\dontrun{
filtered_stream_add_rule(token = your_token,
                         value = "",
                         tag = "")

}

}
