% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenlist.R
\name{flattenlist}
\alias{flattenlist}
\title{This function is a sub function helping make a list flat.}
\usage{
flattenlist(x)
}
\arguments{
\item{x}{list with elements from the twitter API}
}
\value{
a list
}
\description{
This function is a sub function helping make a list flat.
}
\examples{
\dontrun{
}
}
