% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeline.R
\name{get_timelines_v2}
\alias{get_timelines_v2}
\title{This function allows you to collect up to the last 3200 tweets from a users timeline from twitter with the users user_id}
\usage{
get_timelines_v2(
  token = NA,
  user_id = "",
  tweet_fields = "ALL",
  user_fields = "ALL",
  expansions = "ALL",
  place_fields = "ALL",
  poll_fields = "NONE",
  n = 100,
  JSON = FALSE
)
}
\arguments{
\item{token}{string of the bearer token from your personal twitter API access}

\item{user_id}{string of the users id from the user of which you would like to get the timeline from}

\item{tweet_fields}{string which defaults to ALL (no other argument accepted at the moment)}

\item{user_fields}{string which defaults to ALL (no other argument accepted at the moment)}

\item{expansions}{string which defaults to ALL (no other argument accepted at the moment)}

\item{place_fields}{string which defaults to ALL (no other argument accepted at the moment)}

\item{poll_fields}{string which defaults to NONE (no other argument accepted at the moment)}

\item{n}{integer representing the last n tweets to get from the timeline (maximum is 3200 and default is 100)}

\item{JSON}{boolean which defaults to FALSE (no other argument accepted at the moment)}
}
\value{
a data frame
}
\description{
This function allows you to collect up to the last 3200 tweets from a users timeline from twitter with the users user_id
}
\examples{
\dontrun{
users <- get_timelines_v2(token=bearer_token, user_id = "2244994945", n = 100)
}
}
