% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_stream.R
\name{parse_sampled_stream}
\alias{parse_sampled_stream}
\title{Converts Twitter stream data (JSON file) into parsed data frame.}
\usage{
parse_sampled_stream(path)
}
\arguments{
\item{path}{Character, name of JSON file with data collected by
[stream_tweets()].}
}
\value{
a data frame
}
\description{
Converts Twitter stream data (JSON file) into parsed data frame.
}
\examples{
\dontrun{
stream_tweets(timeout = 1, file_name = "stream.json", parse = FALSE)
parse_stream("stream.json")
}

}
