% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_list_prep_i.R
\name{stream_list_prep_i}
\alias{stream_list_prep_i}
\title{This function is a sub function combining listed columns into a string for simpler handling in a data frame}
\usage{
stream_list_prep_i(dat = dat, string = string)
}
\arguments{
\item{dat}{data frame of tweets from stream}

\item{string}{character string with pattern of columns to deal with}
}
\value{
a data frame
}
\description{
This function is a sub function combining listed columns into a string for simpler handling in a data frame
}
