% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_transformer.R
\name{tweets_transformer}
\alias{tweets_transformer}
\title{This function is a sub function transforming many list type columns in the twitter data to a simple column with string of comma delimited entries}
\usage{
tweets_transformer(dt)
}
\arguments{
\item{dt}{data table or data frame}
}
\value{
a data frame or data table
}
\description{
This function is a sub function transforming many list type columns in the twitter data to a simple column with string of comma delimited entries
}
