% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_stream_retrieve_rule.R
\name{filtered_stream_get_rules}
\alias{filtered_stream_get_rules}
\title{This function allows you to collect tweets from the 1 \% sampled stream from twitter}
\usage{
filtered_stream_get_rules(token = "")
}
\arguments{
\item{token}{string of the bearer token from your personal twitter API access}
}
\value{
a list
}
\description{
This function allows you to collect tweets from the 1 \% sampled stream from twitter
}
\examples{
\dontrun{
filtered_stream_get_rules(token = your_token,)

}

}
